+++ Assets/LDtkUnity/Editor/ScriptedImporter/LDtkProjectImporter.cs
@@ -136,7 +136,7 @@ namespace LDtkUnity.Editor
             Profiler.EndSample();
 
             Profiler.BeginSample("SetPixelsPerUnit");
-            SetPixelsPerUnit((int) json.DefaultGridSize); //if for whatever reason (or backwards compatibility), if the ppu is -1 in any capacity
+            SetPixelsPerUnit(json.DefaultGridSize); //if for whatever reason (or backwards compatibility), if the ppu is -1 in any capacity
             Profiler.EndSample();
             
             Profiler.BeginSample("CacheRecentImporter");
@@ -329,12 +329,12 @@ namespace LDtkUnity.Editor
         public TileBase GetTileArtifact(TilesetDefinition tileset, Rect srcPos)
         {
             string relPath = FixRelPath(tileset);
-            return GetArtifactAsset(_artifacts.GetIndexedTile, () => LDtkKeyFormatUtil.GetGetterSpriteOrTileAssetName(srcPos, relPath, (int)tileset.PxHei));
+            return GetArtifactAsset(_artifacts.GetIndexedTile, () => LDtkKeyFormatUtil.GetGetterSpriteOrTileAssetName(srcPos, relPath, tileset.PxHei));
         }
         public Sprite GetSpriteArtifact(TilesetDefinition tileset, Rect srcPos)
         {
             string relPath = FixRelPath(tileset);
-            return GetArtifactAsset(_artifacts.GetIndexedSprite, () => LDtkKeyFormatUtil.GetGetterSpriteOrTileAssetName(srcPos, relPath, (int)tileset.PxHei));
+            return GetArtifactAsset(_artifacts.GetIndexedSprite, () => LDtkKeyFormatUtil.GetGetterSpriteOrTileAssetName(srcPos, relPath, tileset.PxHei));
         }
         public Sprite GetBackgroundArtifact(Level level, int textureHeight)
         {
