+++ Assets/LDtkUnity/Runtime/Data/Schema/Definition/AutoLayerRuleGroup.cs
@@ -23,7 +23,7 @@ namespace LDtkUnity
         public AutoLayerRuleDefinition[] Rules { get; set; }
 
         [DataMember(Name = "uid")]
-        public long Uid { get; set; }
+        public int Uid { get; set; }
 
         [DataMember(Name = "usesWizard")]
         public bool UsesWizard { get; set; }
